Remove deprecated automation code
Fixes #1478
