+++ bin/gpu_usage_plot
@@ -5,6 +5,6 @@ import sys
 import os
 sys.path.append(os.path.abspath('.'))
 
-from util.automation import GPUUsageChart
+from util.gpu_usage import GPUUsageChart
 
 GPUUsageChart(sys.argv[1], sys.argv[2])
